Move NovelAI settings to respective module
